Adding health system.

+++ Assets/Project/Scripts/Game/Player/IPlayer.cs
@@ -4,7 +4,7 @@ namespace game
 {
     public interface IPlayer
     {
-        public void OnGetAttacked(float damage);
+        public void OnGetAttacked(int damage);
         public void OnPathBlocked();
         public void OnPathUnblocked();
         public Vector3 GetPosition();
